Move apollo-server-env mock into a global package
