+++ packages/apollos-apollo-server-env-mock/.npmignore
@@ -1,0 +1,2 @@
+*
+!lib/**
