+++ packages/apollos-apollo-server-env-mock/README.md
@@ -1,0 +1,11 @@
+# `@apollos/config`
+
+> TODO: description
+
+## Usage
+
+```
+const config = require('@apollos/config');
+
+// TODO: DEMONSTRATE API
+```
