+++ packages/apollos-apollo-server-env-mock/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "@apollosproject/apollo-server-env-mock",
+  "version": "0.8.0-alpha.3",
+  "homepage": "https://github.com/ApollosProject/apollos-prototype#readme",
+  "main": "lib/index.js",
+  "scripts": {
+    "build": "NODE_ENV=production rm -rf ./lib && mkdir -p lib && babel ./src -d ./lib"
+  },
+  "devDependencies": {
+    "@apollosproject/babel-preset-apollos": "^0.8.0-alpha.3",
+    "jest-fetch-mock": "^2.0.0"
+  },
+  "license": "SEE LICENSE IN LICENSE.md",
+  "publishConfig": {
+    "access": "public"
+  }
+}
