+++ packages/apollos-apollo-server-env-mock/src/apollo-datasource-mocks/live-stream.js
