+++ packages/apollos-apollo-server-env-mock/src/rock-api-mocks/content-item.js
