+++ packages/apollos-church-api/package.json
@@ -55,6 +55,7 @@
   },
   "devDependencies": {
     "@apollosproject/babel-preset-apollos": "^0.8.0-alpha.3",
+    "@apollosproject/apollo-server-env-mock": "^0.8.0-alpha.3",
     "@babel/cli": "7.1.5",
     "@babel/core": "7.1.6",
     "codecov": "3.0.2",
