--- packages/apollos-data-connector-church-online/__mocks__/apollo-datasource-mocks/index.js
@@ -1,1 +1,0 @@
-export { default as liveStreamLive } from './live-stream';
