--- packages/apollos-data-connector-cloudinary/__mocks__/apollo-datasource-mocks/index.js
@@ -1,1 +1,0 @@
-export { default as Scripture } from './bible';
