+++ packages/apollos-data-connector-cloudinary/package.json
@@ -24,6 +24,7 @@
   },
   "devDependencies": {
     "@apollosproject/babel-preset-apollos": "^0.8.0-alpha.3",
+    "@apollosproject/apollo-server-env-mock": "^0.8.0-alpha.3",
     "jest": "^23.6.0"
   },
   "publishConfig": {
