--- packages/apollos-data-connector-people/__mocks__/rock-api-mocks/index.js
@@ -1,4 +1,0 @@
-export { default as people } from './people';
-export { default as userLogins } from './user-logins';
-export { default as session } from './session';
-export { default as families } from './families';
