--- packages/apollos-data-connector-people/__mocks__/rock-api-mocks/session.js
@@ -1,3 +1,0 @@
-export default () => ({
-  Id: 123,
-});
