--- packages/apollos-data-connector-rock-auth/__mocks__/rock-api-mocks/index.js
@@ -1,3 +1,0 @@
-export { default as people } from './people';
-export { default as userLogins } from './user-logins';
-export { default as session } from './session';
