--- packages/apollos-data-connector-rock-auth/__mocks__/rock-api-mocks/user-logins.js
@@ -1,5 +1,0 @@
-export default () => ({
-  UserName: 'isaac.hardy@newspring.cc',
-  PersonId: 35,
-  Id: 21,
-});
