+++ packages/apollos-data-connector-rock-auth/package.json
@@ -32,6 +32,7 @@
     "@apollosproject/babel-preset-apollos": "^0.8.0-alpha.3",
     "@apollosproject/config": "^0.8.0-alpha.3",
     "@apollosproject/rock-apollo-data-source": "^0.8.0-alpha.3",
+    "@apollosproject/apollo-server-env-mock": "^0.8.0-alpha.3",
     "graphql": "0.13.2"
   },
   "license": "SEE LICENSE IN LICENSE.md",
