--- packages/apollos-data-connector-rock-content/__mocks__/apollo-datasource-mocks/index.js
@@ -1,2 +1,0 @@
-export { default as liveStreamLive } from './live-stream';
-export { default as Scripture } from './bible';
