--- packages/apollos-data-connector-rock-content/__mocks__/apollo-datasource-mocks/live-stream.js
@@ -1,6 +1,0 @@
-export default () => ({
-  meta: { status: 200 },
-  response: {
-    item: { isLive: true, eventStartTime: '2018-08-06T17:00:00Z' },
-  },
-});
