--- packages/apollos-data-connector-rock-content/__mocks__/rock-api-mocks/families.js
@@ -1,117 +1,0 @@
-export default () => [
-  {
-    Campus: {
-      IsSystem: false,
-      Name: 'Main Campus',
-      Description: null,
-      IsActive: true,
-      ShortCode: 'MAIN',
-      Url: null,
-      LocationId: 2,
-      PhoneNumber: null,
-      LeaderPersonAliasId: null,
-      ServiceTimes: null,
-      Order: 0,
-      TimeZoneId: null,
-      CreatedDateTime: null,
-      ModifiedDateTime: null,
-      CreatedByPersonAliasId: null,
-      ModifiedByPersonAliasId: null,
-      Id: 1,
-      Guid: '76882ae3-1ce8-42a6-a2b6-8c0b29cf8cf8',
-      ForeignId: null,
-      ForeignGuid: null,
-      ForeignKey: null,
-    },
-    GroupLocations: [
-      {
-        Location: {
-          ParentLocationId: null,
-          Name: null,
-          IsActive: true,
-          LocationTypeValueId: null,
-          Street1: '6414 S Green St',
-          Street2: null,
-          City: 'Chicago',
-          County: 'Cook',
-          State: 'IL',
-          Country: 'US',
-          PostalCode: '60621-1924',
-          Barcode: '606211924144',
-          AssessorParcelId: null,
-          StandardizeAttemptedDateTime: '2018-10-02T11:11:29.873',
-          StandardizeAttemptedServiceType: 'Smarty Streets',
-          StandardizeAttemptedResult:
-            'RecordType:S; DPV MatchCode:Y; Precision:Zip9',
-          StandardizedDateTime: '2018-10-02T11:11:29.873',
-          GeocodeAttemptedDateTime: '2018-10-02T11:11:29.873',
-          GeocodeAttemptedServiceType: 'Smarty Streets',
-          GeocodeAttemptedResult:
-            'RecordType:S; DPV MatchCode:Y; Precision:Zip9',
-          GeocodedDateTime: '2018-10-02T11:11:29.873',
-          IsGeoPointLocked: null,
-          PrinterDeviceId: null,
-          ImageId: null,
-          SoftRoomThreshold: null,
-          FirmRoomThreshold: null,
-          Latitude: 41.77745,
-          Longitude: -87.64603,
-          Distance: 0,
-          CreatedDateTime: '2018-10-02T11:11:29.953',
-          ModifiedDateTime: '2018-10-02T11:11:29.953',
-          CreatedByPersonAliasId: 106,
-          ModifiedByPersonAliasId: 106,
-          Id: 45,
-          Guid: '571c6044-695e-45c3-bf53-b9728c270298',
-          ForeignId: null,
-          ForeignGuid: null,
-          ForeignKey: null,
-        },
-        GroupId: 373,
-        LocationId: 45,
-        GroupLocationTypeValueId: 19,
-        IsMailingLocation: true,
-        IsMappedLocation: true,
-        GroupMemberPersonAliasId: null,
-        Order: 0,
-        CreatedDateTime: '2018-10-02T11:16:02.487',
-        ModifiedDateTime: '2018-10-02T11:16:02.487',
-        CreatedByPersonAliasId: 106,
-        ModifiedByPersonAliasId: 106,
-        Id: 58,
-        Guid: '91312c3b-f211-4e11-9b5f-321080d22c30',
-        ForeignId: null,
-        ForeignGuid: null,
-        ForeignKey: null,
-      },
-    ],
-    IsSystem: false,
-    ParentGroupId: null,
-    GroupTypeId: 10,
-    CampusId: 1,
-    ScheduleId: null,
-    Name: 'Family',
-    Description: null,
-    IsSecurityRole: false,
-    IsActive: true,
-    Order: 0,
-    AllowGuests: null,
-    WelcomeSystemEmailId: null,
-    ExitSystemEmailId: null,
-    SyncDataViewId: null,
-    AddUserAccountsDuringSync: null,
-    MustMeetRequirementsToAddMember: null,
-    IsPublic: true,
-    GroupCapacity: null,
-    RequiredSignatureDocumentTemplateId: null,
-    CreatedDateTime: '2018-08-09T13:04:43.097',
-    ModifiedDateTime: '2018-10-02T11:15:58.4',
-    CreatedByPersonAliasId: 64,
-    ModifiedByPersonAliasId: 106,
-    Id: 373,
-    Guid: '23534528-6236-4038-980c-39c7c332ead2',
-    ForeignId: null,
-    ForeignGuid: null,
-    ForeignKey: null,
-  },
-];
