--- packages/apollos-data-connector-rock-content/__mocks__/rock-api-mocks/index.js
@@ -1,10 +1,0 @@
-export {
-  default as contentChannelItemAssociation,
-} from './content-channel-item-association';
-export { default as contentChannel } from './content-channel';
-export { default as contentItem } from './content-item';
-export { default as people } from './people';
-export { default as userLogins } from './user-logins';
-export { default as session } from './session';
-export { default as interaction } from './interaction';
-export { default as families } from './families';
